more fun with microsoft's newest.  vc8 has unsigned 64-bit long size_t and spits out warnings when assigning int = container.size().  also uses 64-bit pointers, so pointer arithmetic might need casts too.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@12513 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/AudioMenu.cxx
@@ -102,7 +102,7 @@ AudioMenu::AudioMenu()
   option->update();
   listHUD.push_back(option);
 
-  initNavigation(listHUD, 1, listHUD.size() - 1);
+  initNavigation(listHUD, 1, (int)listHUD.size() - 1);
 }
 
 AudioMenu::~AudioMenu()
@@ -144,7 +144,7 @@ void			AudioMenu::resize(int _width, int _height)
   x = 0.5f * ((float)_width);
   y -= 0.6f * titleHeight;
   const float h = fm.getStrHeight(fontFace, fontSize, " ");
-  const int count = listHUD.size();
+  const int count = (const int)listHUD.size();
   for (i = 1; i < count; i++) {
     listHUD[i]->setFontSize(fontSize);
     listHUD[i]->setPosition(x, y);
